CREATED SOCKETS

+++ Configuration/Server.cpp
@@ -26,7 +26,7 @@ Server  & Server::operator=(const Server & S){
 // 	return server_config;
 // }
 
-int 					Server::GetPort(){return port;}
+size_t 					Server::GetPort(){return port;}
 int 					Server::GetSizeLimit(){return size_limit;}
 std::string 			Server::GetHost(){return host;}
 std::string 			Server::GetServerName(){return server_name;}
@@ -34,7 +34,7 @@ std::string 			Server::GetErrorPage(){return error_page;}
 std::vector<Location>	Server::GetLocation(){return location;}
 
 //SETTERS;
-void Server::SetPort(int port){this->port = port;}
+void Server::SetPort(size_t port){this->port = port;}
 void Server::SetSizeLimit(int size_limit){this->size_limit = size_limit;}
 void Server::SetHost(std::string host){this->host = host;}
 void Server::SetServerName(std::string server_name){this->server_name = server_name;}
